+++ hadoop-ozone/tools/src/test/java/org/apache/hadoop/ozone/debug/TestCheckNative.java
@@ -15,14 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.hadoop.ozone.checknative;
+package org.apache.hadoop.ozone.debug;
 
 import static org.apache.hadoop.hdds.utils.NativeConstants.ROCKS_TOOLS_NATIVE_LIBRARY_NAME;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.apache.hadoop.hdds.utils.IOUtils;
-import org.apache.hadoop.ozone.debug.OzoneDebug;
-import org.apache.hadoop.ozone.shell.checknative.CheckNative;
 import org.apache.ozone.test.GenericTestUtils;
 import org.apache.ozone.test.tag.Native;
 import org.junit.jupiter.api.AfterEach;
